Capture full paths to autotools at project generation time (#2)
Xcode IDE based builds run in a sanitized environment which
does not include directories such as /opt/homebrew/bin. When
the cmake generated build scripts depend on commands outside
the sanitized environment the build fails. The find_program
function is used to capture the absolute paths to executables
at project generation time which allows the builds to run in
the environment provieded by Xcode.